Updated name for IDE
The SmartThings IDE allows you to link it directly to Github if the files are properly named. This pull is to rename it so that it links to SmartThings. This will allow people to be more easily notified of any changes you might make in the code, and to very quickly update their code to the most recent version.